Moved auxillary output mode to config values packet and added manual control packet

+++ OASMan_ESP32/src/manifoldSaveData.cpp
@@ -397,7 +397,7 @@ AuxillaryOutputMode getauxillaryOutputMode() {
 AuxillaryOutputModeTimeUnit getauxillaryOutputModeTimeUnit() {
     return (AuxillaryOutputModeTimeUnit)_SaveData.auxillaryOutputPreference.auxillaryOutputModeTimeUnit.get().i;
 }
-uint32_t getauxillaryOutputTime() {
+uint16_t getauxillaryOutputTime() {
     return _SaveData.auxillaryOutputPreference.auxillaryOutputTime.get().i;
 }
